Review 02 Changes
Removed IBaseMouse / BaseMouse
+ Attached Mouse is just a decorator for Mouse. Should not have any
common denominator via Interface with Mouse

Added IAttachedMouse
+ Decorator AttachedMouse should have it's own interface

Removed IBaseKeyboard / BaseKeyboard
+ AttachedKeyboard is just a decorator Class for Keyboard and should
have no common denominator via an Interface

Added IAttachedKeyboard
+ Decorator AttachedKeyboard should have it's own interface

Moved content from IBaseMouse to IMouse
Moved content from BaseMouse to Mouse
Moved content from IBaseKeyboard to IKeyboard
Moved content from BaseKeyboard to Keyboard
