+++ src/TestStack.White/InputDevices/AttachedMouse.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Windows;
 using TestStack.White.UIItems;
 using TestStack.White.UIItems.Actions;
@@ -9,7 +8,7 @@ namespace TestStack.White.InputDevices
     /// Any operation performed using the mouse would wait till the window is busy after this operation. Before any operation is 
     /// performed the window, from which it was retreived, is brought to focus if it is not.
     /// </summary>
-    public class AttachedMouse : IBaseMouse
+    public class AttachedMouse : IAttachedMouse
     {
         private readonly IActionListener actionListener;
         private readonly IMouse mouse;
@@ -21,7 +20,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.Location"/>
+        /// Implements <see cref="IAttachedMouse.Location"/>
         /// </summary>
         public virtual Point Location
         {
@@ -30,7 +29,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.Cursor"/>
+        /// Implements <see cref="IAttachedMouse.Cursor"/>
         /// </summary>
         public virtual MouseCursor Cursor
         {
@@ -38,7 +37,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.Click(MouseButton)"/>
+        /// Implements <see cref="IAttachedMouse.Click(MouseButton)"/>
         /// </summary>
         public virtual void Click(MouseButton mouseButton)
         {
@@ -46,7 +45,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.Click(MouseButton, Point)"/>
+        /// Implements <see cref="IAttachedMouse.Click(MouseButton, Point)"/>
         /// </summary>
         public virtual void Click(MouseButton mouseButton, Point point)
         {
@@ -54,7 +53,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DoubleClick(MouseButton)"/>
+        /// Implements <see cref="IAttachedMouse.DoubleClick(MouseButton)"/>
         /// </summary>
         public virtual void DoubleClick(MouseButton mouseButton)
         {
@@ -62,7 +61,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DoubleClick(MouseButton, Point)"/>
+        /// Implements <see cref="IAttachedMouse.DoubleClick(MouseButton, Point)"/>
         /// </summary>
         public virtual void DoubleClick(MouseButton mouseButton, Point point)
         {
@@ -70,7 +69,7 @@ namespace TestStack.White.InputDevices
         }
         
         /// <summary>
-        /// Implements <see cref="IBaseMouse.LeftClick()"/>
+        /// Implements <see cref="IAttachedMouse.LeftClick()"/>
         /// </summary
         public virtual void LeftClick()
         {
@@ -79,7 +78,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.LeftClick(Point)"/>
+        /// Implements <see cref="IAttachedMouse.LeftClick(Point)"/>
         /// </summary
         public virtual void LeftClick(Point point)
         {
@@ -87,7 +86,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.LeftDoubleClick()"/>
+        /// Implements <see cref="IAttachedMouse.LeftDoubleClick()"/>
         /// </summary
         public virtual void LeftDoubleClick()
         {
@@ -96,7 +95,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.LeftDoubleClick(Point)"/>
+        /// Implements <see cref="IAttachedMouse.LeftDoubleClick(Point)"/>
         /// </summary
         public virtual void LeftDoubleClick(Point point)
         {
@@ -105,7 +104,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.RightClick(Point)"/>
+        /// Implements <see cref="IAttachedMouse.RightClick(Point)"/>
         /// </summary
         public virtual void RightClick()
         {
@@ -114,7 +113,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.RightClick(Point)"/>
+        /// Implements <see cref="IAttachedMouse.RightClick(Point)"/>
         /// </summary
         public virtual void RightClick(Point point)
         {
@@ -122,7 +121,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragAndDrop(IUIItem, IUIItem)"/>
+        /// Implements <see cref="IAttachedMouse.DragAndDrop(IUIItem, IUIItem)"/>
         /// </summary
         public virtual void DragAndDrop(IUIItem draggedItem, IUIItem dropItem)
         {
@@ -131,7 +130,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragAndDrop(MouseButton, IUIItem, IUIItem)"/>
+        /// Implements <see cref="IAttachedMouse.DragAndDrop(MouseButton, IUIItem, IUIItem)"/>
         /// </summary
         public virtual void DragAndDrop(MouseButton mouseButton, IUIItem draggedItem, IUIItem dropItem)
         {
@@ -140,7 +139,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragAndDrop(IUIItem, Point, IUIItem, Point)"/>
+        /// Implements <see cref="IAttachedMouse.DragAndDrop(IUIItem, Point, IUIItem, Point)"/>
         /// </summary
         public virtual void DragAndDrop(IUIItem draggedItem, Point startPosition, IUIItem dropItem, Point endPosition)
         {
@@ -149,7 +148,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragAndDrop(MouseButton, IUIItem, Point, IUIItem, Point)"/>
+        /// Implements <see cref="IAttachedMouse.DragAndDrop(MouseButton, IUIItem, Point, IUIItem, Point)"/>
         /// </summary
         public virtual void DragAndDrop(MouseButton mouseButton, IUIItem draggedItem, Point startPosition, IUIItem dropItem, Point endPosition)
         {
@@ -158,7 +157,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragHorizontally(IUIItem, int)"/>
+        /// Implements <see cref="IAttachedMouse.DragHorizontally(IUIItem, int)"/>
         /// </summary
         public virtual void DragHorizontally(IUIItem uiItem, int distance)
         {
@@ -167,7 +166,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragHorizontally(MouseButton, IUIItem, int)"/>
+        /// Implements <see cref="IAttachedMouse.DragHorizontally(MouseButton, IUIItem, int)"/>
         /// </summary
         public virtual void DragHorizontally(MouseButton mouseButton, IUIItem uiItem, int distance)
         {
@@ -176,7 +175,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragVertically(IUIItem, int)"/>
+        /// Implements <see cref="IAttachedMouse.DragVertically(IUIItem, int)"/>
         /// </summary
         public virtual void DragVertically(IUIItem uiItem, int distance)
         {
@@ -185,7 +184,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.DragVertically(MouseButton, IUIItem, int)"/>
+        /// Implements <see cref="IAttachedMouse.DragVertically(MouseButton, IUIItem, int)"/>
         /// </summary
         public virtual void DragVertically(MouseButton mouseButton, IUIItem uiItem, int distance)
         {
@@ -194,7 +193,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.MoveOut()"/>
+        /// Implements <see cref="IAttachedMouse.MoveOut()"/>
         /// </summary
         public virtual void MoveOut()
         {
@@ -203,7 +202,7 @@ namespace TestStack.White.InputDevices
         }
 
         /// <summary>
-        /// Implements <see cref="IBaseMouse.Move(Point)"/>
+        /// Implements <see cref="IAttachedMouse.Move(Point)"/>
         /// </summary
         public virtual void Move(Point position)
         {
