+++ src/TestStack.White/InputDevices/IKeyboard.cs
@@ -3,14 +3,26 @@ using TestStack.White.WindowsAPI;
 
 namespace TestStack.White.InputDevices
 {
-    public interface IKeyboard : IBaseKeyboard
+    public interface IKeyboard
     {
+        /// <summary>
+        /// Enter a string
+        /// </summary>
+        /// <param name="keysToType">String to enter</param>
+        void Enter(string keysToType);
+
         /// <summary>
         /// Send a strig to a specific <see cref="IActionListener"/>
         /// </summary>
         /// <param name="keysToType">String to send</param>
         /// <param name="actionListener"><see cref="IActionListener"/> to receive the string</param>
         void Send(string keysToType, IActionListener actionListener);
+
+        /// <summary>
+        /// Press special Keys
+        /// </summary>
+        /// <param name="key"><see cref="KeyboardInput.SpecialKeys"/> to press</param>
+        void PressSpecialKey(KeyboardInput.SpecialKeys key);
 
         /// <summary>
         /// Press special Keys
@@ -23,6 +35,12 @@ namespace TestStack.White.InputDevices
         /// Hold a special key
         /// </summary>
         /// <param name="key"><see cref="KeyboardInput.SpecialKeys"/> to hold</param>
+        void HoldKey(KeyboardInput.SpecialKeys key);
+
+        /// <summary>
+        /// Hold a special key
+        /// </summary>
+        /// <param name="key"><see cref="KeyboardInput.SpecialKeys"/> to hold</param>
         /// <param name="actionListener"><see cref="IActionListener"/> to receive the string</param>
         void HoldKey(KeyboardInput.SpecialKeys key, IActionListener actionListener);
 
@@ -30,8 +48,29 @@ namespace TestStack.White.InputDevices
         /// Leave a pressed Key
         /// </summary>
         /// <param name="key">Key to leave</param>
+        void LeaveKey(KeyboardInput.SpecialKeys key);
+
+        /// <summary>
+        /// Leave a pressed Key
+        /// </summary>
+        /// <param name="key">Key to leave</param>
         /// <param name="actionListener"><see cref="IActionListener"/> to receive the string</param>
         void LeaveKey(KeyboardInput.SpecialKeys key, IActionListener actionListener);
+
+        /// <summary>
+        /// Get / Set the CapsLock
+        /// </summary>
+        bool CapsLockOn { get; set; }
+
+        /// <summary>
+        /// List of Held Keys
+        /// </summary>
+        KeyboardInput.SpecialKeys[] HeldKeys { get; }
+
+        /// <summary>
+        /// Leave all Keys
+        /// </summary>
+        void LeaveAllKeys();
 
         /// <summary>
         /// Perform an Action on the Action Listener
