+++ src/TestStack.White/TestStack.White.csproj
@@ -99,10 +99,12 @@
     <Compile Include="Drawing\ScreenRectangle.cs" />
     <Compile Include="InputDevices\BareMetalKeyboard.cs" />
     <Compile Include="InputDevices\BareMetalMouse.cs" />
-    <Compile Include="InputDevices\BaseKeyboard.cs" />
-    <Compile Include="InputDevices\BaseMouse.cs" />
-    <Compile Include="InputDevices\IBaseKeyboard.cs" />
-    <Compile Include="InputDevices\IBaseMouse.cs" />
+    <Compile Include="InputDevices\Keyboard.cs" />
+    <Compile Include="InputDevices\IAttachedKeyboard.cs" />
+    <Compile Include="InputDevices\Mouse.cs" />
+    <Compile Include="InputDevices\IKeyboard.cs" />
+    <Compile Include="InputDevices\IAttachedMouse.cs" />
+    <Compile Include="InputDevices\IMouse.cs" />
     <Compile Include="InputDevices\Input64.cs" />
     <Compile Include="InputDevices\Native.cs" />
     <Compile Include="InputDevices\MouseButton.cs" />
@@ -149,11 +151,7 @@
     <Compile Include="HorizontalSpan.cs" />
     <Compile Include="InputDevices\AttachedKeyboard.cs" />
     <Compile Include="InputDevices\AttachedMouse.cs" />
-    <Compile Include="InputDevices\IKeyboard.cs" />
-    <Compile Include="InputDevices\IMouse.cs" />
     <Compile Include="InputDevices\InputDeviceException.cs" />
-    <Compile Include="InputDevices\Keyboard.cs" />
-    <Compile Include="InputDevices\Mouse.cs" />
     <Compile Include="InputDevices\MouseCursor.cs" />
     <Compile Include="Interceptors\CoreInterceptContext.cs" />
     <Compile Include="Interceptors\ScrollInterceptor.cs" />
